# backend/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Aluguel de Carros API
  description: |
    API REST do serviço de aluguel de carros (Micronaut).

    **Autenticação:** a maioria das rotas exige header `Authorization: Bearer <JWT>`.

    **Rotas públicas (sem token):**
    - `POST /auth/login`
    - `POST /clientes`
    - `POST /bancos`
    - `POST /empresas`
    - `GET /swagger-ui`
    - `GET /swagger-ui/swagger.yml`
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/aluguelcarros/v1
    description: Ambiente local
...
paths:
...
  /clientes:
    post:
      tags: [Clientes]
      summary: Cadastrar cliente
      description: |
        Rota pública.

        O cliente deve informar entre 1 e 3 entidades empregadoras.
      operationId: cadastrarCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequest'
            example:
              rg: "MG1234567"
              cpf: "12345678901"
              nome: "João da Silva"
              endereco: "Rua A, 100"
              profissao: "Analista"
              senha: "senha123"
              entidadesEmpregadoras:
                - nomeEmpresa: "Empresa XPTO"
                  cnpj: "11222333000181"
                  rendimento: 4500.0
      responses:
        '200':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '400':
          description: Erro de validação
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
    get:
      tags: [Clientes]
      summary: Listar clientes
      operationId: listarClientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteResponse'
        '401':
          description: Não autorizado
  /clientes/{id}:
    parameters:
      - $ref: '#/components/parameters/ClienteId'
    get:
      tags: [Clientes]
      summary: Buscar cliente por id
      operationId: buscarClientePorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '401':
          description: Não autorizado
        '404':
          description: Cliente não encontrado
    put:
      tags: [Clientes]
      summary: Atualizar cliente
      operationId: atualizarCliente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequest'
      responses:
        '200':
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Cliente não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Clientes]
      summary: Remover cliente
      operationId: removerCliente
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cliente removido
        '401':
          description: Não autorizado
        '404':
          description: Cliente não encontrado
...
  /empresas:
    post:
      tags: [Empresas]
      summary: Cadastrar empresa
      description: |
        Rota pública.

        A empresa autentica por CNPJ e usa o CNPJ como login interno.
      operationId: cadastrarEmpresa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpresaRequest'
            example:
              razaoSocial: "Empresa XPTO LTDA"
              cnpj: "99887766000155"
              ramoDeAtividade: "Tecnologia"
              senha: "senhaEmpresa123"
      responses:
        '200':
          description: Empresa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpresaResponse'
        '400':
          description: Erro de validação
        '422':
          description: Regra de negócio violada
    get:
      tags: [Empresas]
      summary: Listar empresas
      operationId: listarEmpresas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmpresaResponse'
        '401':
          description: Não autorizado
  /empresas/{id}:
    parameters:
      - $ref: '#/components/parameters/EmpresaId'
    get:
      tags: [Empresas]
      summary: Buscar empresa por id
      operationId: buscarEmpresaPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Empresa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpresaResponse'
        '401':
          description: Não autorizado
        '404':
          description: Empresa não encontrada
    put:
      tags: [Empresas]
      summary: Atualizar empresa
      operationId: atualizarEmpresa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpresaRequest'
      responses:
        '200':
          description: Empresa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpresaResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Empresa não encontrada
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Empresas]
      summary: Remover empresa
      operationId: removerEmpresa
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Empresa removida
        '401':
          description: Não autorizado
        '404':
          description: Empresa não encontrada
...
  /pareceres:
    post:
      tags: [Pareceres]
      summary: Emitir parecer para pedido
      description: |
        Regras de negócio:
        - só pode emitir parecer para pedido com status `EM_ANALISE`
        - ao emitir parecer, o status do pedido é atualizado para `APROVADO` ou `REJEITADO`
      operationId: cadastrarParecer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParecerRequest'
            example:
              pedidoId: 1
              agenteTipo: "BANCO"
              agenteId: 1
              resultado: "APROVADO"
              justificativa: "Renda comprovada e histórico favorável."
              dataEmissao: "2026-04-14"
      responses:
        '200':
          description: Parecer emitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParecerResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Entidade relacionada não encontrada
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
    get:
      tags: [Pareceres]
      summary: Listar pareceres
      operationId: listarPareceres
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pareceres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParecerResponse'
        '401':
          description: Não autorizado
  /pareceres/{id}:
    parameters:
      - $ref: '#/components/parameters/ParecerId'
    get:
      tags: [Pareceres]
      summary: Buscar parecer por id
      operationId: buscarParecerPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Parecer encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParecerResponse'
        '401':
          description: Não autorizado
        '404':
          description: Parecer não encontrado
    put:
      tags: [Pareceres]
      summary: Atualizar parecer
      operationId: atualizarParecer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParecerRequest'
      responses:
        '200':
          description: Parecer atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParecerResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Parecer não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Pareceres]
      summary: Remover parecer
      operationId: removerParecer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Parecer removido
        '401':
          description: Não autorizado
        '404':
          description: Parecer não encontrado
  /contratos:
    post:
      tags: [Contratos]
      summary: Cadastrar contrato
      description: |
        Regras de negócio:
        - contrato só pode ser criado para pedido com status `APROVADO`
        - cada pedido pode possuir apenas um contrato
      operationId: cadastrarContrato
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratoRequest'
            example:
              pedidoId: 1
              tipoPropriedade: "CLIENTE"
              dataAssinatura: "2026-04-14"
      responses:
        '200':
          description: Contrato criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Entidade relacionada não encontrada
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
    get:
      tags: [Contratos]
      summary: Listar contratos
      operationId: listarContratos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de contratos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContratoResponse'
        '401':
          description: Não autorizado
  /contratos/{id}:
    parameters:
      - $ref: '#/components/parameters/ContratoId'
    get:
      tags: [Contratos]
      summary: Buscar contrato por id
      operationId: buscarContratoPorId
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contrato encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Contrato não encontrado
    put:
      tags: [Contratos]
      summary: Atualizar contrato
      operationId: atualizarContrato
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratoRequest'
      responses:
        '200':
          description: Contrato atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '400':
          description: Erro de validação
        '401':
          description: Não autorizado
        '404':
          description: Contrato não encontrado
        '422':
          description: Regra de negócio violada
    delete:
      tags: [Contratos]
      summary: Remover contrato
      operationId: removerContrato
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Contrato removido
        '401':
          description: Não autorizado
        '404':
          description: Contrato não encontrado
  /contratos/{id}/executar:
    parameters:
      - $ref: '#/components/parameters/ContratoId'
    post:
      tags: [Contratos]
      summary: Executar contrato
      description: |
        Executa o contrato e atualiza o status do automóvel do pedido:
        - `CLIENTE` -> `ALUGADO`
        - `EMPRESA` ou `BANCO` -> `VINCULADO`
      operationId: executarContrato
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contrato executado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratoResponse'
        '401':
          description: Não autorizado
        '404':
          description: Contrato não encontrado
        '422':
          description: Regra de negócio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensagemErro'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado em `POST /auth/login`
  parameters:
    ClienteId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    EmpresaId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ParecerId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ContratoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    EntidadeEmpregadoraRequest:
      type: object
      required: [nomeEmpresa, cnpj, rendimento]
      properties:
        nomeEmpresa:
          type: string
        cnpj:
          type: string
        rendimento:
          type: number
          format: double
    EntidadeEmpregadoraResponse:
      type: object
      properties:
        nomeEmpresa:
          type: string
        cnpj:
          type: string
        rendimento:
          type: number
          format: double
    ClienteRequest:
      type: object
      required: [rg, cpf, nome, endereco, profissao, senha, entidadesEmpregadoras]
      properties:
        rg:
          type: string
        cpf:
          type: string
        nome:
          type: string
        endereco:
          type: string
        profissao:
          type: string
        senha:
          type: string
          format: password
        entidadesEmpregadoras:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: '#/components/schemas/EntidadeEmpregadoraRequest'
    ClienteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rg:
          type: string
        cpf:
          type: string
        nome:
          type: string
        endereco:
          type: string
        profissao:
          type: string
        entidadesEmpregadoras:
          type: array
          items:
            $ref: '#/components/schemas/EntidadeEmpregadoraResponse'
...
    EmpresaRequest:
      type: object
      required: [razaoSocial, cnpj, ramoDeAtividade, senha]
      properties:
        razaoSocial:
          type: string
        cnpj:
          type: string
        ramoDeAtividade:
          type: string
        senha:
          type: string
          format: password
    EmpresaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        razaoSocial:
          type: string
        cnpj:
          type: string
        ramoDeAtividade:
          type: string
...
    AgenteTipo:
      type: string
      enum: [BANCO, EMPRESA]
    ParecerResultado:
      type: string
      enum: [APROVADO, REJEITADO]
    ParecerRequest:
      type: object
      required: [pedidoId, agenteTipo, agenteId, resultado, justificativa, dataEmissao]
      properties:
        pedidoId:
          type: integer
          format: int64
        agenteTipo:
          $ref: '#/components/schemas/AgenteTipo'
        agenteId:
          type: integer
          format: int64
        resultado:
          $ref: '#/components/schemas/ParecerResultado'
        justificativa:
          type: string
        dataEmissao:
          type: string
          format: date
    ParecerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pedidoId:
          type: integer
          format: int64
        agenteTipo:
          $ref: '#/components/schemas/AgenteTipo'
        agenteId:
          type: integer
          format: int64
        resultado:
          $ref: '#/components/schemas/ParecerResultado'
        justificativa:
          type: string
        dataEmissao:
          type: string
          format: date
    TipoPropriedade:
      type: string
      enum: [CLIENTE, EMPRESA, BANCO]
    ContratoRequest:
      type: object
      required: [pedidoId, tipoPropriedade, dataAssinatura]
      properties:
        pedidoId:
          type: integer
          format: int64
        tipoPropriedade:
          $ref: '#/components/schemas/TipoPropriedade'
        dataAssinatura:
          type: string
          format: date
    ContratoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pedidoId:
          type: integer
          format: int64
        automovelId:
          type: integer
          format: int64
        tipoPropriedade:
          $ref: '#/components/schemas/TipoPropriedade'
        dataAssinatura:
          type: string
          format: date
...
    MensagemErro:
      type: object
      properties:
        mensagem:
          type: string